Improve performance of RBT (#41165)

+++ lib/dns/include/dns/db.h
@@ -194,7 +194,7 @@ typedef struct dns_dbmethods {
 				   dns_rdataset_t *rdataset,
 				   dns_rdataset_t *sigrdataset);
 	isc_result_t	(*setcachestats)(dns_db_t *db, isc_stats_t *stats);
-	unsigned int	(*hashsize)(dns_db_t *db);
+	size_t		(*hashsize)(dns_db_t *db);
 } dns_dbmethods_t;
 
 typedef isc_result_t
@@ -1380,7 +1380,7 @@ dns_db_nodecount(dns_db_t *db);
  * \li	The number of nodes in the database
  */
 
-unsigned int
+size_t
 dns_db_hashsize(dns_db_t *db);
 /*%<
  * For database implementations using a hash table, report the
@@ -1392,7 +1392,7 @@ dns_db_hashsize(dns_db_t *db);
  *
  * Returns:
  * \li	The number of buckets in the database's hash table, or
- *      ISC_R_NOTIMPLEMENTED.
+ *      0 if not implemented.
  */
 
 void
